refactored xrefTable to its own component
