+++ ui/src/app/tables.scss
@@ -81,6 +81,10 @@ table {
     max-width: unset;
   }
 
+  th.bordered, td.bordered {
+    border-right: 1px solid $aleph-table-border-color;
+  }
+
   td.entity {
     font-weight: bold;
     max-width: unset;
